Updated header to be custom
- Removed HeaderAccount that is a partial copy of HeaderAccount from pagopa/mui-italia
- Implemented CustomLandingHeaderAccount that is a custom header with only the necessary props
- updated Layout to use CustomLandingHeaderAccount
