+++ src/components/index.ts
@@ -1,7 +1,7 @@
 export * from './Dtd'
 export * from './EnvSwitch'
 export * from './Goals'
-export * from './HeaderAccount'
+export * from './CustomLandingHeaderAccount'
 export * from './HeadMeta'
 export * from './LawSnippets'
 export * from './LineChartSection'
